Explicitly cast daddr_t's to long long's when printing them since daddr_t is not long long on all archs.  (They happen to be long's on 64-bit arch's and gcc considers that significant enough to warn about it.)  These should probably be uintmax_t but I didn't feel like adding all the extra includes.

+++ sys/dev/vinum/vinumdaemon.c
@@ -114,7 +114,7 @@ vinum_daemon(void)
 			rq->bp->b_iocmd == BIO_READ ? "Read" : "Write",
 			major(rq->bp->b_dev),
 			minor(rq->bp->b_dev),
-			rq->bp->b_blkno,
+			(long long)rq->bp->b_blkno,
 			rq->bp->b_bcount);
 		}
 		recover_io(request->info.rq);		    /* the failed request */
